refactor: review - useLazyInterval to own file
